Move utility functions that don't depend on server-side logic to `utils`
