+++ src/adapter-nextjs/page-props/category.ts
@@ -1,8 +1,8 @@
 import type { ExtraStoryFields, Story } from '@prezly/sdk';
 import type { GetServerSidePropsContext, GetServerSidePropsResult } from 'next';
 
-import { DEFAULT_PAGE_SIZE } from '../../data-fetching';
 import type { PaginationProps } from '../../infinite-loading/types';
+import { DEFAULT_PAGE_SIZE } from '../../utils';
 import { getNewsroomServerSideProps } from '../getNewsroomServerSideProps';
 import { processRequest } from '../processRequest';
