+++ src/data-fetching/api/index.ts
@@ -1,2 +1,1 @@
 export * from './getPrezlyApi';
-export * from './languages';
