+++ src/data-fetching/index.ts
@@ -1,13 +1,3 @@
 export * from './api';
-export * from './boilerplate';
-export * from './constants';
-export * from './getAssetsUrl';
-export * from './getDataRequestLink';
-export * from './getLocalizedCategoryData';
-export * from './getNewsroomLogoUrl';
-export * from './getPrivacyPortalUrl';
-export * from './getStoryPublicationDate';
-export * from './getUploadcareGroupUrl';
-export * from './isEmbargoStory';
 export * from './lib';
 export * from './types';
