+++ src/data-fetching/lib/index.ts
@@ -1,2 +1,3 @@
+export { getAlgoliaSettings } from './getAlgoliaSettings';
 export { getEnvVariables } from './getEnvVariables';
 export { isSdkError } from './isSdkError';
