+++ src/index.ts
@@ -5,3 +5,4 @@ export * from './infinite-loading';
 export * from './intl';
 export * from './newsroom-context';
 export * from './types';
+export * from './utils';
