+++ src/intl/index.ts
@@ -1,4 +1,4 @@
-export * from './lang';
+export * from './languages';
 export * from './locale';
 export { default as localeConfig } from './localeConfig';
 export * from './localeObject';
