+++ src/intl/languages.ts
@@ -1,6 +1,44 @@
-import type { NewsroomLanguageSettings, Story } from '@prezly/sdk';
+import type { CultureRef, NewsroomLanguageSettings, Story } from '@prezly/sdk';
 
-import { DUMMY_DEFAULT_LOCALE, LocaleObject } from '../../intl';
+import { DUMMY_DEFAULT_LOCALE } from './locale';
+import { LocaleObject } from './localeObject';
+
+function isOnlyCulture(culture: CultureRef, languages: NewsroomLanguageSettings[]): boolean {
+    const numberOfLanguages = languages.filter(
+        ({ locale: { language_code } }) => language_code === culture.language_code,
+    ).length;
+
+    return numberOfLanguages === 1;
+}
+
+/**
+ * @returns the display name of the locale in its native language
+ *
+ * If there's only one culture used in a specific language,
+ * we strip the culture name completely.
+ *
+ * Examples:
+ *  - English (Global), Spanish (Spain)
+ *  - -> English, Spanish
+ *  - English (Global), English (UK), Spanish (Spain)
+ *  - -> English (Global), English (UK), Spanish
+ */
+export function getLanguageDisplayName(
+    language: NewsroomLanguageSettings,
+    languages: NewsroomLanguageSettings[],
+): string {
+    const { locale } = language;
+
+    if (isOnlyCulture(locale, languages)) {
+        const cultureNameIndex = locale.native_name.indexOf('(');
+
+        if (cultureNameIndex !== -1) {
+            return locale.native_name.slice(0, cultureNameIndex - 1);
+        }
+    }
+
+    return locale.native_name;
+}
 
 /**
  * @returns Language set as default in the Newsroom Settings.
